+++ plom/baseMessenger.py
@@ -49,8 +49,11 @@ from plom.plom_exceptions import (
 
 # We can support earlier servers by special-case code, so
 # define an allow-list of versions we support.
+# TODO: Issue #3717: introduce new reassign API
 Supported_Server_API_Versions = [
     int(Plom_Legacy_Server_API_Version),
+    # 112,
+    # 113,  # introduced /MK/tasks/{code}/reassign/{username}
     int(Plom_API_Version),
 ]
 
@@ -206,9 +209,28 @@ class BaseMessenger:
         return self.whoami()
 
     def is_legacy_server(self) -> bool | None:
+        """Check if the server is the older legacy server.
+
+        Returns:
+            True/False, or None if we're not connected.
+        """
         if self.get_server_API_version() is None:
             return None
         return self.get_server_API_version() == int(Plom_Legacy_Server_API_Version)
+
+    def is_server_api_less_than(self, api_number: int) -> bool | None:
+        """Check if the server API is strictly less than a value.
+
+        Args:
+            api_number: what value to compare to.
+
+        Returns:
+            True/False, or None if we're not connected.
+        """
+        ver = self.get_server_API_version()
+        if ver is None:
+            return None
+        return int(ver) < api_number
 
     @property
     def server(self) -> str: